// src/model/api/surveyUnitMetadata.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * queen-api
 * API for Queen/Stromae
 * OpenAPI spec version: 4.3.10-SNAPSHOT
 */
import type { Logos } from './logos'
import type { MetadataVariable } from './metadataVariable'
import type { SchemaPersonalization } from './schema.personalization/schemaPersonalization'
import type { SurveyUnitMetadataContext } from './surveyUnitMetadataContext'

export interface SurveyUnitMetadata {
  context: SurveyUnitMetadataContext
  label: string
  logos?: Logos
  objectives: string
  personalization?: SchemaPersonalization
  variables?: MetadataVariable[]
}